refactor(home): update weather feature dependency to kocolor-specific module
This commit updates the home feature module to point to the local `kocolor` weather implementation instead of the global feature module. This involves updating the Gradle project dependency and refactoring the corresponding package imports.

- **`build.gradle.kts`**:
    - Updated the project dependency from `:features:weather` to `:applications:kocolor:features:weather`.
- **`HomeScreen.kt` & `HomeViewModel.kt`**:
    - Refactored imports for `LayeredWeatherInfoIcon`, `LayeredWeatherUiState`, and `LayeredWeatherCondition` to use the new `com.zoewave.probase.kocolor.features.weather` package path.
