+++ applications/kocolor/apps/mobile/features/home/build.gradle.kts
@@ -21,7 +21,7 @@ dependencies {
     implementation(project(":features:ai:configuration"))
     implementation(project(":features:graphics"))
     implementation(project(":features:health:core"))
-    implementation(project(":features:weather"))
+    implementation(project(":applications:kocolor:features:weather"))
     implementation(project(":core:network"))
     implementation(project(":core:data"))
     implementation(project(":core:model"))
